requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
dnspython==2.0.0
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-vision==0.42.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mysql-connector-python==8.0.19
networkx==2.4
numba==0.48.0
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
passlib==1.7.2
Pillow==7.0.0
protobuf==3.6.1
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
validate-email==1.3
Werkzeug==1.0.0
